move current diff to llmdiff
