+++ packages/core/src/github.ts
@@ -20,7 +20,7 @@ import { shellRemoveAsciiColors } from "./shell"
 import { isGlobMatch } from "./glob"
 import { fetchText } from "./fetch"
 import { concurrentLimit } from "./concurrency"
-import { createDiff, llmifyDiff } from "./diff"
+import { createDiff, llmifyDiff } from "./llmdiff"
 import { JSON5TryParse } from "./json5"
 import { link } from "./mkmd"
 import { LanguageModelInfo } from "./server/messages"
