+++ packages/core/src/parsers.ts
@@ -29,7 +29,7 @@ import { resolveFileContent } from "./file"
 import { resolveTokenEncoder } from "./encoders"
 import { mustacheRender } from "./mustache"
 import { jinjaRender } from "./jinja"
-import { createDiff, llmifyDiff } from "./diff"
+import { createDiff, llmifyDiff } from "./llmdiff"
 import { tidyData } from "./tidy"
 import { hash } from "./crypto"
 import { GROQEvaluate } from "./groq"
