+++ packages/core/src/promptdom.ts
@@ -37,7 +37,7 @@ import { ChatCompletionMessageParam } from "./chattypes"
 import { resolveTokenEncoder } from "./encoders"
 import { expandFileOrWorkspaceFiles } from "./fs"
 import { interpolateVariables } from "./mustache"
-import { createDiff } from "./diff"
+import { createDiff } from "./llmdiff"
 import { promptyParse } from "./prompty"
 import { jinjaRenderChatMessage } from "./jinja"
 import { runtimeHost } from "./host"
